Version 1.0 - Initial full release
Updated P commands to support 3Pyyy to set the temporary brightness.
This allows a command to change the brightness, enabling more support
for the Stealth system.  3Pyyy is not saved in EEPROM so is safe to use
in all cases.
Calling 3P0 will restore the brightness value to whatever previous
setting that was, either internal global value, or the POT.

Also updated the max brightness to be 175 to preserve LED life.